+++ apps/react/app/App.tsx
@@ -7,8 +7,8 @@ import { listener } from './listener'
 import styles from './page.module.css'
 
 const SPACE_ID = 'us_sp_123456'
-const ENVIRONMENT_ID = 'us_env_O60vqdol'
-const PUBLISHABLE_KEY = 'pk_3626978e5f52480085a37844166d347a'
+const ENVIRONMENT_ID = 'us_env_123456'
+const PUBLISHABLE_KEY = 'pk_123456'
 
 const spaceProps = {
   environmentId: ENVIRONMENT_ID,
@@ -43,10 +43,6 @@ function App() {
     loading: <LoadingComponent />,
     exitPrimaryButtonText: 'CLOSE!',
     exitSecondaryButtonText: 'KEEP IT!',
-    closeSpace: {
-      operation: 'contacts:submit',
-      onClose: () => setShowSpace(false),
-    },
   })
 
   const SimpleSpace = ({
@@ -74,10 +70,6 @@ function App() {
           record.addInfo('lastName', 'Welcome to the Rock fam')
         }
         return record
-      },
-      closeSpace: {
-        operation: 'simpleSubmitAction',
-        onClose: () => setShowSpace(false),
       },
     })
     return portal
@@ -113,11 +105,7 @@ function App() {
           {showSimplified === true ? 'Close' : 'Open'} Pre-loaded
         </button>
       </div>
-      {showSpace && (
-        <div className={styles.spaceWrapper}>
           <Space />
-        </div>
-      )}
       {showSimplified && (
         <div>
           <SimpleSpace setShowSpace={setShowSimplified} />