+++ packages/react/src/hooks/usePortal.tsx
@@ -17,7 +17,9 @@ import { recordHook } from '@flatfile/plugin-record-hook'
 import { IReactSimpleOnboarding } from '../types/IReactSimpleOnboarding'
 import api from '@flatfile/api'
 
-export const usePortal = (props: IReactSimpleOnboarding): JSX.Element => {
+export const usePortal = (
+  props: IReactSimpleOnboarding
+): JSX.Element | null => {
   const { errorTitle, loading: LoadingElement, apiUrl } = props
   const [initError, setInitError] = useState<string>()
   const [state, setState] = useState<State>({
@@ -26,6 +28,7 @@ export const usePortal = (props: IReactSimpleOnboarding): JSX.Element => {
     spaceUrl: '',
   })
   const [flatfileOptions, setFlatfileOptions] = useState(props)
+  const [closeInstance, setCloseInstance] = useState<boolean>(false)
 
   const { localSpaceId, spaceUrl, accessTokenLocal } = state
   const onSubmitSettings = { ...DefaultSubmitSettings, ...props.submitSettings }
@@ -48,8 +51,10 @@ export const usePortal = (props: IReactSimpleOnboarding): JSX.Element => {
                       record: FlatfileRecord,
                       event: FlatfileEvent | undefined
                     ) => {
-                      // @ts-ignore - something weird with the `data` prop and the types upstream in the packages being declared in different places, but overall this is fine
-                      return config.onRecordHook(record, event)
+                      return (
+                        config.onRecordHook &&
+                        config.onRecordHook(record, event)
+                      )
                     }
                   )
                 )
@@ -155,6 +160,10 @@ export const usePortal = (props: IReactSimpleOnboarding): JSX.Element => {
     return errorElement
   }
 
+  if (closeInstance) {
+    return null
+  }
+
   if (localSpaceId && spaceUrl && accessTokenLocal) {
     return (
       <Space
@@ -162,6 +171,7 @@ export const usePortal = (props: IReactSimpleOnboarding): JSX.Element => {
         spaceId={localSpaceId}
         spaceUrl={spaceUrl}
         accessToken={accessTokenLocal}
+        handleCloseInstance={() => setCloseInstance(true)}
         {...flatfileOptions}
       />
     )