+++ packages/listener-driver-pubsub/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @flatfile/listener-driver-pubsub
 
+## 2.0.3
+
+### Patch Changes
+
+- Updated dependencies [56388f0]
+  - @flatfile/listener@1.0.0
+
+## 2.0.2
+
+### Patch Changes
+
+- Updated dependencies [2ed22cb]
+  - @flatfile/listener@0.4.2
+
 ## 2.0.1
 
 ### Patch Changes