+++ packages/listener/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @flatfile/listener
 
+## 1.0.0
+
+### Major Changes
+
+- 56388f0: Update package.json to use exports nested entrypoints
+
+## 0.4.2
+
+### Patch Changes
+
+- 2ed22cb: Fix to help separate types from conflicting peer dependencies
+
 ## 0.4.1
 
 ### Patch Changes