+++ packages/v2-shims/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @flatfile/v2-shims
 
+## 1.1.1
+
+### Patch Changes
+
+- d3e68f1: Update types for all packages that reference the record-hook plugin
+
+## 1.1.0
+
+### Minor Changes
+
+- 56388f0: Update package.json to use exports nested entrypoints
+
+### Patch Changes
+
+- Updated dependencies [56388f0]
+  - @flatfile/listener@1.0.0
+
 ## 1.0.1
 
 ### Patch Changes